api/requirements.txt
# MUSE/api/requirements.txt

fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
# elevenlabs 버전을 명확히 지정 (예: 최신 버전 확인 후)
elevenlabs==1.3.0 # 또는 확인된 최신 안정 버전
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0